refactor: 유저 infra 계층 패키지 구조 정리
