+++ src/main/java/kr/hhplus/be/server/user/domain/port/UserRepository.java
@@ -1,6 +1,7 @@
-package kr.hhplus.be.server.user.domain;
+package kr.hhplus.be.server.user.domain.port;
 
 import java.util.Optional;
+import kr.hhplus.be.server.user.domain.User;
 
 public interface UserRepository {
     Optional<User> findById(long userId);
